Moved parsers and commands
