+++ Assets/Scripts/UI/InGameUI/DevConsole/CommandsAndParsers.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: a440edacdca79f14f9e8e047af853cbd
+timeCreated: 1492213416
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
