+++ Assets/Scripts/UI/InGameUI/DevConsole/DevConsole.cs
@@ -605,246 +605,6 @@ namespace DeveloperConsole
             Parsers.Add(target, new ParseDelegate(method));
         }
 
-        #region AttributeImplementations
-
-        /// <summary>
-        /// Logs all the tags in a nice format.
-        /// </summary>
-        [Command("System", description = "Logs all the tags used currently by the system", title = "AllTags")]
-        public static void AllTags()
-        {
-            Log("All the tags: ", "green");
-            Log(string.Join(", ", CommandArray().SelectMany(x => x.Tags).Select(x => x.Trim()).Distinct().ToArray()));
-        }
-
-        /// <summary>
-        /// Just returns help dependent on each command.
-        /// </summary>
-        /// <param name="tag"> The tag to search for.  If "" then do all tags. </param>
-        [Command("System", description = "Returns information on all commands.  Can take in a parameter as a tag to search for all commands with that tag", title = "Help")]
-        public static void Help(string tag = "")
-        {
-            Log("-- Help --", "green");
-
-            string text = string.Empty;
-
-            CommandBase[] consoleCommands = CommandArray(tag);
-
-            for (int i = 0; i < consoleCommands.Length; i++)
-            {
-                text += "\n<color=orange>" + consoleCommands[i].Title + GetParameters(consoleCommands[i]) + "</color>" + (consoleCommands[i].DescriptiveText == null ? string.Empty : " //" + consoleCommands[i].DescriptiveText);
-            }
-
-            Log(text);
-
-            Log("\n<color=orange>Note:</color> If the function has no parameters you <color=red> don't</color> need to use the parameter modifier.");
-            Log("<color=orange>Note:</color> You <color=red>don't</color> need to use the trailing parameter modifier either");
-            Log("You can use constants to replace common parameters (they are case insensitive but require ' ' around them):");
-            Log("- 'Center' (or 'Centre') is the position of the center/centre of the map.");
-            Log("- 'MousePos' is the position of the mouse");
-            Log("- 'TimeScale' is the current time scale");
-            Log("- 'Pi' is 3.141...");
-            Log("- 'E' is 2.718...");
-        }
-
-        /// <summary>
-        /// Clears the text area and history.
-        /// </summary>
-        [Command("System", description = "Clears the developer console", title = "Clear")]
-        public static void Clear()
-        {
-            ClearHistory();
-            Text textObj = TextObject();
-
-            if (textObj != null)
-            {
-                TextObject().text = "\n<color=#7CFC00>Clear Successful :D</color>\n";
-            }
-        }
-
-        /// <summary>
-        /// Parse an integer 16 from arguments given.
-        /// </summary>
-        /// <param name="args"> The string version of the end result. </param>
-        /// <returns> Null if arguments are wrong else the converted arg in object form. </returns>
-        [Parser(typeof(short))]
-        public static object HandleInt16(string args)
-        {
-            short outValue;
-            return short.TryParse(args, out outValue) ? (object)outValue : null;
-        }
-
-        /// <summary>
-        /// Parse an integer 32 from arguments given.
-        /// </summary>
-        /// <param name="args"> The string version of the end result. </param>
-        /// <returns> Null if arguments are wrong else the converted arg in object form. </returns>
-        [Parser(typeof(int))]
-        public static object HandleInt32(string args)
-        {
-            int outValue;
-            return int.TryParse(args, out outValue) ? (object)outValue : null;
-        }
-
-        /// <summary>
-        /// Parse an integer 64 from arguments given.
-        /// </summary>
-        /// <param name="args"> The string version of the end result. </param>
-        /// <returns> Null if arguments are wrong else the converted arg in object form. </returns>
-        [Parser(typeof(long))]
-        public static object HandleInt64(string args)
-        {
-            long outValue;
-            return long.TryParse(args, out outValue) ? (object)outValue : null;
-        }
-
-        /// <summary>
-        /// Parse an unsigned integer 16 from arguments given.
-        /// </summary>
-        /// <param name="args"> The string version of the end result. </param>
-        /// <returns> Null if arguments are wrong else the converted arg in object form. </returns>
-        [Parser(typeof(ushort))]
-        public static object HandleUInt16(string args)
-        {
-            ushort outValue;
-            return ushort.TryParse(args, out outValue) ? (object)outValue : null;
-        }
-
-        /// <summary>
-        /// Parse an unsigned integer 32 from arguments given.
-        /// </summary>
-        /// <param name="args"> The string version of the end result. </param>
-        /// <returns> Null if arguments are wrong else the converted arg in object form. </returns>
-        [Parser(typeof(uint))]
-        public static object HandleUInt32(string args)
-        {
-            uint outValue;
-            return uint.TryParse(args, out outValue) ? (object)outValue : null;
-        }
-
-        /// <summary>
-        /// Parse an unsigned integer 64 from arguments given.
-        /// </summary>
-        /// <param name="args"> The string version of the end result. </param>
-        /// <returns> Null if arguments are wrong else the converted arg in object form. </returns>
-        [Parser(typeof(ulong))]
-        public static object HandleUInt64(string args)
-        {
-            ulong outValue;
-            return ulong.TryParse(args, out outValue) ? (object)outValue : null;
-        }
-
-        /// <summary>
-        /// Parse a float from arguments given.
-        /// </summary>
-        /// <param name="args"> The string version of the end result. </param>
-        /// <returns> Null if arguments are wrong else the converted arg in object form. </returns>
-        [Parser(typeof(float))]
-        public static object HandleFloat(string args)
-        {
-            float outValue;
-            return float.TryParse(args, out outValue) ? (object)outValue : null;
-        }
-
-        /// <summary>
-        /// Parse a double from arguments given.
-        /// </summary>
-        /// <param name="args"> The string version of the end result. </param>
-        /// <returns> Null if arguments are wrong else the converted arg in object form. </returns>
-        [Parser(typeof(double))]
-        public static object HandleDouble(string args)
-        {
-            double outValue;
-            return double.TryParse(args, out outValue) ? (object)outValue : null;
-        }
-
-        /// <summary>
-        /// Parse a decimal from arguments given.
-        /// </summary>
-        /// <param name="args"> The string version of the end result. </param>
-        /// <returns> Null if arguments are wrong else the converted arg in object form. </returns>
-        [Parser(typeof(decimal))]
-        public static object HandleDecimal(string args)
-        {
-            decimal outValue;
-            return decimal.TryParse(args, out outValue) ? (object)outValue : null;
-        }
-
-        /// <summary>
-        /// Parse a boolean from arguments given.
-        /// </summary>
-        /// <param name="args"> The string version of the end result. </param>
-        /// <returns> Null if arguments are wrong else the converted arg in object form. </returns>
-        [Parser(typeof(bool))]
-        public static object HandleBoolean(string args)
-        {
-            bool outValue;
-            return bool.TryParse(args, out outValue) ? (object)outValue : null;
-        }
-
-        /// <summary>
-        /// Parse a string from arguments given.
-        /// </summary>
-        /// <param name="args"> The string version of the end result. </param>
-        /// <returns> Null if arguments are wrong else the converted arg in object form. </returns>
-        [Parser(typeof(string))]
-        public static object HandleString(string args)
-        {
-            return args;
-        }
-
-        /// <summary>
-        /// Parse an float from arguments given.
-        /// </summary>
-        /// <param name="args"> The string version of the end result. </param>
-        /// <returns> Null if arguments are wrong else the converted arg in object form. </returns>
-        [Parser(typeof(char))]
-        public static object HandleChar(string args)
-        {
-            return args.Length >= 1 ? (object)args.ToCharArray()[0] : null;
-        }
-
-        /// <summary>
-        /// Parse a vector2 from arguments given.
-        /// </summary>
-        /// <param name="args"> The string version of the end result. </param>
-        /// <returns> Null if arguments are wrong else the converted arg in object form. </returns>
-        [Parser(typeof(Vector2))]
-        public static object HandleVector2(string args)
-        {
-            string[] coordinates = args.Split(',');
-            float x, y;
-            return coordinates.Length >= 2 && float.TryParse(coordinates[0], out x) && float.TryParse(coordinates[1], out y) ? (object)new Vector2(x, y) : null;
-        }
-
-        /// <summary>
-        /// Parse a vector3 from arguments given.
-        /// </summary>
-        /// <param name="args"> The string version of the end result. </param>
-        /// <returns> Null if arguments are wrong else the converted arg in object form. </returns>
-        [Parser(typeof(Vector3))]
-        public static object HandleVector3(string args)
-        {
-            string[] coordinates = args.Split(',');
-            float x, y, z;
-            return coordinates.Length >= 3 && float.TryParse(coordinates[0], out x) && float.TryParse(coordinates[1], out y) && float.TryParse(coordinates[2], out z) ? (object)new Vector3(x, y, z) : null;
-        }
-
-        /// <summary>
-        /// Parse a vector4 from arguments given.
-        /// </summary>
-        /// <param name="args"> The string version of the end result. </param>
-        /// <returns> Null if arguments are wrong else the converted arg in object form. </returns>
-        [Parser(typeof(Vector4))]
-        public static object HandleVector4(string args)
-        {
-            string[] coordinates = args.Split(',');
-            float x, y, z, w;
-            return coordinates.Length >= 4 && float.TryParse(coordinates[0], out x) && float.TryParse(coordinates[1], out y) && float.TryParse(coordinates[2], out z) && float.TryParse(coordinates[3], out w) ? (object)new Vector4(x, y, z, w) : null;
-        }
-
-        #endregion
-
         /// <summary>
         /// Button delegate action to handle command.
         /// </summary>
