+++ Assets/StreamingAssets/Data/ConsoleCommands.xml
@@ -8,11 +8,9 @@
   <!--You set parameters to the format Type VariableName,...-->
   <!--E.g. UnityEngine.Vector3 MyVector, Int32 MyPrimitiveInt, Single MyPrimitiveFloat -->
   <!--Note: You can chooes not to include the namespace if its a system namespace (like primitives i.e. Int32, Boolean...)-->
-  <ConsoleCommand Title="SetFontSize" FunctionName="SetFontSize" Description="Sets the font size, must be between 10 and 20 (inclusive)" Parameters="Int32 size" DetailedDescription="This command sets the font size of the console.  It takes a single parameter which is a whole number (integer) and has to be between 10 and 20 (inclusive).  The default is 14." Tags="Console, UI" DefaultValue="14"></ConsoleCommand>
   <ConsoleCommand Title="SetTimeStamp" FunctionName="SetTimeStamp" Description="Sets the time stamp" Parameters="Boolean on" Tags="Console, UI" DefaultValue=""></ConsoleCommand>
   <ConsoleCommand Title="CameraPos" FunctionName="ChangeCameraPosition" Description="Changes the camera position to the one provided" Parameters="UnityEngine.Vector3 NewPos" Tags="Camera" DefaultValue="'Center'"></ConsoleCommand>
   <ConsoleCommand Title="Run_LUA" FunctionName="Run_LUA" Description="Runs the string as LUA" Parameters="String luaText" Tags="LUA, Modding, Compile" DefaultValue=""></ConsoleCommand>
-  <ConsoleCommand Title="SetText" FunctionName="SetText" Description="Sets the console text" Parameters="String text" Tags="Console, UI" DefaultValue=""></ConsoleCommand>
   <ConsoleCommand Title="SetCharacterHealth" FunctionName="SetCharacterHealth" Description="Sets the character's health to the value provided" Parameters="String name, Single health" Tags="Character, Health"></ConsoleCommand>
   <ConsoleCommand Title="DamageCharacter" FunctionName="DamageCharacter" Description="Damages the character by the amount provided" Parameters="String name, Single amount" Tags="Character, Health"></ConsoleCommand>
   <ConsoleCommand Title="CharacterHealthSystemSet" FunctionName="CharacterHealthSystemSet" Description="Sets the character's health system" Parameters="String name, Single hp, Boolean overheal, Boolean healable, Boolean invincible, Boolean revivable" Tags="Character, Health"></ConsoleCommand>
@@ -31,9 +29,7 @@
   <ConsoleCommand Title="GetThermalDiffusivity" FunctionName="GetThermalDiffusivity" Description="Gets the thermal diffusivity at the position" Parameters="UnityEngine.Vector3 pos" Tags="Temperature" DefaultValue="{Center}"></ConsoleCommand>
   <ConsoleCommand Title="FloodFillRoom" FunctionName="FloodFillRoomAt" Description="Floods fills the room at the position provided" Parameters="UnityEngine.Vector3 pos" Tags="Room" DefaultValue="'Center'"></ConsoleCommand>
   <ConsoleCommand Title="GetAllRoomIDs" FunctionName="GetAllRoomIDs" Description="Logs all the room ids" Parameters="" Tags="Room" DefaultValue=""></ConsoleCommand>
-  <ConsoleCommand Title="Exit" FunctionName="Exit" Description="Exits the console" Parameters="" Tags="System" DefaultValue=""></ConsoleCommand>
-  <ConsoleCommand Title="DevMode" FunctionName="DevMode" Description="Floods fills the room at the position provided" Parameters="Boolean isOn" Tags="Developer" DefaultValue="true"></ConsoleCommand>
-  <ConsoleCommand Title="Status" FunctionName="Status" Description="Returns current status of game" Parameters="" Tags="System" DefaultValue=""></ConsoleCommand>
+  <ConsoleCommand Title="DevMode" FunctionName="DevMode" Description="Turns dev mode on/off." Parameters="Boolean isOn" Tags="Developer" DefaultValue="true"></ConsoleCommand>
   <ConsoleCommand Title="NewCharacter" FunctionName="NewCharacter" Description="Creates a new character at the position provided" Parameters="UnityEngine.Vector3 pos, string name" Tags="Character" DefaultValue="{Center}, &quot;&quot;"></ConsoleCommand>
   <ConsoleCommand Title="DoBuild" FunctionName="DoBuild" DetailedDescription="Does build mode using the furniture/floor/whatever type provided at position pos.  The options for build mode are: FLOOR = 0, ROOMBEHAVIOUR= 1, FURNITURE= 2, UTILITY = 3, and DECONSTRUCT= 4" Description="Builds the object provided at the position passed in" Parameters="Int32 buildMode, string type, UnityEngine.Vector3 pos" Tags="Build" DefaultValue=""></ConsoleCommand>
   <ConsoleCommand Title="InvalidateTileGraph" FunctionName="InvalidateTileGraph" Description="Invalidates the tile graph" Parameters="" Tags="Check" DefaultValue=""></ConsoleCommand>
